\name{internal}

\alias{setpars}
\alias{getpars}
\alias{pmod}
\alias{predpmod}
\alias{obj}
\alias{predict.pmod}
\alias{print.pmod}
\alias{summary.pmod}
\alias{reorderDrift}


\title{
Functions for internal use. 
}


\description{
Functions for specifying, printing and altering the submodels for 
each of the parameters. 
}

\author{
Ingmar Visser
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

