\name{bh08}
\alias{bh08}
\docType{data}
\title{
Example data from Brown & Heathcote (2008). 
}

\description{

Example data from Brown & Heathcote (2008); response times and accuracies
for three difficulty levels.

}

\usage{data(bh08)}
\format{
  A data frame with 3000 observations on the following 3 variables.
  \describe{
    \item{\code{acc}}{response accuracy}
    \item{\code{rt}}{response times}
    \item{\code{diff}}{difficulty, a factor with levels \code{easy} \code{medium} \code{diff}}
  }
}
\details{
Data retrieved from: http://www.newcl.org/members/chris/fitLBA.zip
}
\source{
Brown & Heathcote, 2008
}
\references{
Brown & Heathcote, 2008
}
\examples{
data(bh08)
}
\keyword{datasets}
