\name{core}

\alias{fptpdf}
\alias{fptcdf}
\alias{n1PDF}
\alias{getPars}
\alias{setPars}

\title{

Core functions to compute the probability density function, cumulative
distribution function, and defective distribution of the LBA (see Brown &
Heathcote 2008), and some other utility functions.

}

\description{
For internal use only.

}

\references{

These functions are retrieved from:
http://www.newcl.org/members/chris/fitLBA.zip

}

\author{

The functions fptpdf, fptcdf and n1PDF are provided by Brown, Heathcote and
Donkin and are retrieved from:
http://www.newcl.org/members/chris/fitLBA.zip

}
