\name{core}

\alias{fptpdf}
\alias{fptcdf}
\alias{n1PDF}

\title{

Core functions to compute the probability density function, cumulative
distribution function, and defective distribution of the LBA (see Brown and
Heathcote, 2008).

}

\description{
For internal use only.
}

\references{

Original code was extracted in 2015 here: http://www.newcl.org/members/chris/fitLBA.zip

}

\author{
Ingmar Visser adapted these functions, to make them fully vectorized in all
relevant arguments.  Original functions fptpdf, fptcdf and n1PDF are
provided by Brown, Heathcote and Donkin and were retrieved in 2015 from this 
website: http://www.newcl.org/members/chris/fitLBA.zip

}
