\name{core}

\alias{fptpdf}
\alias{fptcdf}
\alias{n1PDF}

\title{
Core functions to compute the probability density function,
cumulative distribution function, and defective distribution
of the LBA (see Brown & Heathcote 2008). 
}
\description{
For internal use only.
}

\references{
These functions are retrieved from: http://www.newcl.org/members/chris/fitLBA.zip
}

\author{
These functions are provided by Brown, Heathcote and Donkin, 
}
