\name{glba-package}

\alias{glba-package}
\alias{glba}

\docType{package}

\title{Fit LBA models with explanatory variables.}

\description{
Fit LBA models with explanatory variables on the parameters.
}

\details{
\tabular{ll}{
Package: \tab glba\cr
Type: \tab Package\cr
Version: \tab 0.2 \cr
Date: \tab 2015-02-11\cr
License: \tab GPL \cr
}
See ?lba for examples.
}

\author{
Ingmar Visser (with functions adapted from Brown, Heathcote and Donkin).
Maintainer: Ingmar Visser <i.visser@uva.nl>
}

\references{
Brown, S. D., and Heathcote, A. (2008). The simplest complete model of choice response time: linear ballistic accumulation. Cognitive psychology, 57(3), 153-178.
}
\keyword{package}
\keyword{statmodel}

\examples{
# To be added later
}
