\name{glba-package}

\alias{glba-package}
\alias{glba}

\docType{package}

\title{Fit LBA models with explanatory variables.}

\description{
Fit LBA models with explanatory variables on the parameters.
}

\details{
\tabular{ll}{
Package: \tab glba\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2014-07-18\cr
License: \tab GPL \cr
}
See ?lba for examples.
}

\author{
Ingmar Visser (with function from Brown, Heathcote and Donkin)
Maintainer: Ingmar Visser <i.visser@uva.nl>
}

\references{
Brown & Heathcote, 2008.
}

\keyword{package}
\keyword{statmodel}

\examples{
# To be added later
}
