\name{internal}

\alias{setPars}
\alias{getPars}

\alias{pmod}
\alias{predpmod}

\alias{predict.pmod}
\alias{print.pmod}
\alias{summary.pmod}

\alias{obj}
\alias{reorderDrift}


\title{
Utility functions for internal use. 
}


\description{
Functions for specifying, printing and altering the submodels for 
each of the parameters. 
}

\author{
Ingmar Visser
}

